Move the loading view into a separate file.
